Refactor: 앤드포인트작성
