--- investement/src/main/java/com/example/investment/financialRatio/controller/dto/FinancialRatioDTO.java
@@ -1,14 +1,0 @@
-package com.example.investment.financialRatio.controller.dto;
-
-public record FinancialRatioDTO(
-        String stockAccountingYearMonth,
-        String grossMarginRatio,
-        String businessProfitRate,
-        String netInterestRate,
-        String roeValue,
-        String earningsPerShare,
-        String salesPerShare,
-        String bookValuePerShare,
-        String reserveRate,
-        String liabilityRate
-) {}
