+++ investement/src/main/java/com/example/investment/home/fluctuation/controller/FluctuationController.java
@@ -1,8 +1,8 @@
 package com.example.investment.home.fluctuation.controller;
 
-
 import com.example.investment.home.fluctuation.controller.dto.response.FluctuationDTO;
 import com.example.investment.home.fluctuation.service.FluctuationService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,12 +12,11 @@ import java.io.IOException;
 import java.util.List;
 
 @RestController
-@RequestMapping("api/home/fluctuation")
+@RequestMapping("home/fluctuation")
 public class FluctuationController {
 
     private final FluctuationService fluctuationService;
 
-    @Autowired
     public FluctuationController(FluctuationService fluctuationService) {
         this.fluctuationService = fluctuationService;
     }
@@ -26,6 +25,4 @@ public class FluctuationController {
     public List<FluctuationDTO> getFluctuation() throws IOException {
         return fluctuationService.getFluctuation();
     }
-
-
 }
