+++ investement/src/main/java/com/example/investment/home/fluctuation/infrastructor/FluctuationParser.java
@@ -1,8 +1,10 @@
 package com.example.investment.home.fluctuation.infrastructor;
 
 import com.example.investment.home.fluctuation.controller.dto.response.FluctuationDTO;
+
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
@@ -13,39 +15,40 @@ import java.util.List;
 @Component
 public class FluctuationParser {
 
-        private final ObjectMapper objectMapper;
+    private final ObjectMapper objectMapper;
 
-        public FluctuationParser(final ObjectMapper objectMapper) {
-            this.objectMapper = objectMapper;
+    public FluctuationParser(final ObjectMapper objectMapper) {
+        this.objectMapper = objectMapper;
+    }
+
+    public List<FluctuationDTO> getFluctuation(String responseBody) throws IOException {
+
+        JsonNode rootNode = objectMapper.readTree(responseBody);
+        JsonNode items = rootNode.path("output");
+
+        List<FluctuationDTO> fluctuationDTOList = new ArrayList<>();
+        Iterator<JsonNode> elements = items.elements();
+        int count = 0;
+
+        while (elements.hasNext() && count < 5) {
+
+            JsonNode fluctuationItem = elements.next();
+
+            String stockName = fluctuationItem.path("hts_kor_isnm").asText();
+
+            int rank = Integer.parseInt(fluctuationItem.path("data_rank").asText());
+            int currentPrice = Integer.parseInt(fluctuationItem.path("stck_prpr").asText());
+            int prevChangePrice = Integer.parseInt(fluctuationItem.path("prdy_vrss").asText());
+
+            String prevSign = fluctuationItem.path("prdy_vrss_sign").asText();
+            Double prevChangeRate = Double.valueOf(fluctuationItem.path("prdy_ctrt").asText());
+
+            fluctuationDTOList.add(new FluctuationDTO(stockName, rank, currentPrice, prevChangePrice, prevSign, prevChangeRate));
+            count++;
         }
 
-        public List<FluctuationDTO> getFluctuation(String responseBody) throws IOException {
-            System.out.println("ResponseBody: " +responseBody);
-            JsonNode rootNode = objectMapper.readTree(responseBody);
-            System.out.println("rootNode"+rootNode);
-            JsonNode items = rootNode.path("output");
-            System.out.println("items"+items.toString());
-            List<FluctuationDTO> fluctuationDTOList = new ArrayList<>();
-            Iterator<JsonNode> elements = items.elements();
-            System.out.println(elements);
-            int count = 0;
+        return fluctuationDTOList;
+    }
 
-            while (elements.hasNext() && count < 5) {
-                JsonNode fluctuationItem = elements.next();
-
-                String stockName = fluctuationItem.path("hts_kor_isnm").asText();
-                int rank = Integer.parseInt(fluctuationItem.path("data_rank").asText());
-                int currentPrice = Integer.parseInt(fluctuationItem.path("stck_prpr").asText());
-                int prevChangePrice = Integer.parseInt(fluctuationItem.path("prdy_vrss").asText());
-                String prevSign = fluctuationItem.path("prdy_vrss_sign").asText();
-                Double prevChangeRate = Double.valueOf(fluctuationItem.path("prdy_ctrt").asText());
-
-                fluctuationDTOList.add(new FluctuationDTO(stockName, rank, currentPrice, prevChangePrice, prevSign, prevChangeRate));
-                count++;
-            }
-
-            System.out.println(fluctuationDTOList);
-            return fluctuationDTOList;
-        }
 }
