+++ investement/src/main/java/com/example/investment/home/fluctuation/service/FluctuationService.java
@@ -4,6 +4,8 @@ package com.example.investment.home.fluctuation.service;
 import com.example.investment.home.fluctuation.controller.dto.response.FluctuationDTO;
 import com.example.investment.home.fluctuation.infrastructor.FluctuationParser;
 import com.example.investment.home.fluctuation.service.client.FluctuationDataFetcher;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,9 +16,12 @@ import java.util.List;
 @Service
 @Transactional
 public class FluctuationService {
+
     private final FluctuationDataFetcher fluctuationDataFetcher;
+
     private final FluctuationParser fluctuationParser;
 
+    @Autowired
     public FluctuationService(FluctuationDataFetcher fluctuationDataFetcher, FluctuationParser fluctuationParser) {
         this.fluctuationDataFetcher = fluctuationDataFetcher;
         this.fluctuationParser=fluctuationParser;
