+++ investement/src/main/java/com/example/investment/home/index/controller/dto/KOSDAQResponse.java
@@ -1,0 +1,5 @@
+package com.example.investment.home.index.controller.dto;
+
+public record KOSDAQResponse(String indexName, String indexValue, String fluctuationRate) {
+
+}
