+++ investement/src/main/java/com/example/investment/home/tradingVolume/controller/dto/TradingVolumeDTO.java
@@ -1,0 +1,11 @@
+package com.example.investment.home.tradingVolume.controller.dto;
+
+public record TradingVolumeDTO(
+        String stockName,
+        String rank,
+        String currentPrice,
+        String totalVolume,
+        String prevVolume,
+        String volumeChangeRate
+) {
+}
