+++ investement/src/main/java/com/example/investment/home/tradingVolume/service/client/TradingVolumeFetcher.java
@@ -1,0 +1,50 @@
+package com.example.investment.home.tradingVolume.service.client;
+
+import org.springframework.beans.factory.annotation.Value;
+
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
+
+import org.springframework.stereotype.Component;
+
+import org.springframework.web.client.RestTemplate;
+
+@Component
+public class TradingVolumeFetcher {
+
+    private final RestTemplate restTemplate;
+
+    public TradingVolumeFetcher(final RestTemplate restTemplate) {
+        this.restTemplate = restTemplate;
+    }
+
+    @Value("${api.tr_id}")
+    private String trId;
+
+    @Value("${api.app_secret}")
+    private String appSecret;
+
+    @Value("${api.app_key}")
+    private String appKey;
+
+    @Value("${api.access_token}")
+    private String accessToken;
+
+    public ResponseEntity<String> fetchTradingVolumeData() {
+        String url = "https://openapi.koreainvestment.com:9443/uapi/domestic-stock/v1/quotations/volume-rank?FID_COND_MRKT_DIV_CODE=J&FID_COND_SCR_DIV_CODE=20171&FID_INPUT_ISCD=0000&FID_DIV_CLS_CODE=0&FID_BLNG_CLS_CODE=0&FID_TRGT_CLS_CODE=111111111&FID_TRGT_EXLS_CLS_CODE=0000000000&FID_INPUT_PRICE_1=0&FID_INPUT_PRICE_2=0&FID_VOL_CNT=0&FID_INPUT_DATE_1=0";
+        HttpHeaders headers = setHeader();
+        HttpEntity<String> entity = new HttpEntity<>(headers);
+        return restTemplate.exchange(url, HttpMethod.GET, entity, String.class);
+    }
+
+    private HttpHeaders setHeader() {
+        HttpHeaders headers = new HttpHeaders();
+        headers.set("tr_id", trId);
+        headers.set("appsecret", appSecret);
+        headers.set("appkey", appKey);
+        headers.set("Authorization", "Bearer " + accessToken);
+        return headers;
+    }
+}
