--- investement/src/main/java/com/example/investment/index/controller/dto/KOSPIResponse.java
@@ -1,5 +1,0 @@
-package com.example.investment.index.controller.dto;
-
-public record KOSPIResponse(String indexName, String indexValue, String fluctuationRate) {
-
-}
