--- investement/src/main/java/com/example/investment/news/controller/NewsController.java
@@ -1,31 +1,0 @@
-package com.example.investment.news.controller;
-
-import com.example.investment.news.controller.dto.NewsResponse;
-
-import com.example.investment.news.service.NewsService;
-
-import org.json.JSONException;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.io.IOException;
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/news")
-public class NewsController {
-
-    private final NewsService newsService;
-
-    public NewsController(final NewsService newsService) {
-        this.newsService = newsService;
-    }
-
-    @GetMapping
-    public List<NewsResponse> getNews(@RequestParam String keyword) throws JSONException, IOException {
-        return newsService.getNews(keyword);
-    }
-}
