--- investement/src/main/java/com/example/investment/news/controller/dto/NewsResponse.java
@@ -1,5 +1,0 @@
-package com.example.investment.news.controller.dto;
-
-public record NewsResponse(String title, String link) {
-
-}
