+++ investement/src/main/java/com/example/investment/search/financialRatio/controller/dto/FinancialRatioDTO.java
@@ -1,0 +1,14 @@
+package com.example.investment.search.financialRatio.controller.dto;
+
+public record FinancialRatioDTO(
+        String stockAccountingYearMonth,
+        String grossMarginRatio,
+        String businessProfitRate,
+        String netInterestRate,
+        String roeValue,
+        String earningsPerShare,
+        String salesPerShare,
+        String bookValuePerShare,
+        String reserveRate,
+        String liabilityRate
+) {}
