+++ investement/src/main/java/com/example/investment/search/news/controller/dto/NewsResponse.java
@@ -1,0 +1,5 @@
+package com.example.investment.search.news.controller.dto;
+
+public record NewsResponse(String title, String link) {
+
+}
