+++ investement/src/main/java/com/example/investment/search/searchStocks/controller/SearchStocksController.java
@@ -1,0 +1,13 @@
+package com.example.investment.search.searchStocks.controller;
+
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/search/searchStocks")
+public class SearchStocksController {
+
+
+
+}
