+++ investement/src/main/java/com/example/investment/search/searchStocks/controller/dto/SearchStocksDTO.java
@@ -1,0 +1,4 @@
+package com.example.investment.search.searchStocks.controller.dto;
+
+public class SearchStocksDTO {
+}
