--- investement/src/main/java/com/example/investment/searchStocks/controller/SearchStocksController.java
@@ -1,13 +1,0 @@
-package com.example.investment.searchStocks.controller;
-
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/searchStocks")
-public class SearchStocksController {
-
-
-
-}
