--- investement/src/main/java/com/example/investment/searchStocks/controller/dto/SearchStocksDTO.java
@@ -1,4 +1,0 @@
-package com.example.investment.searchStocks.controller.dto;
-
-public class SearchStocksDTO {
-}
