--- investement/src/main/java/com/example/investment/searchStocks/service/client/SearchStocksDataFetcher.java
@@ -1,47 +1,0 @@
-package com.example.investment.searchStocks.service.client;
-
-import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
-
-@Component
-public class SearchStocksDataFetcher {
-
-    @Value("${api.tr_id}")
-    private String trId;
-
-    @Value("${api.app_secret}")
-    private String appSecret;
-
-    @Value("${api.app_key}")
-    private String appKey;
-
-    @Value("${api.access_token}")
-    private String accessToken;
-
-    private final RestTemplate restTemplate;
-
-    public SearchStocksDataFetcher(final RestTemplate restTemplate) {
-        this.restTemplate = restTemplate;
-    }
-
-    public JSONObject SearchStocksData(String fidInputIscd) {
-        String url = "https://openapi.koreainvestment.com:9443/uapi/domestic-stock/v1/quotations/inquire-price?FID_COND_MRKT_DIV_CODE=J&FID_INPUT_ISCD=" + fidInputIscd;
-        HttpHeaders headers = new HttpHeaders();
-        headers.set("tr_id", trId);
-        headers.set("appsecret", appSecret);
-        headers.set("appkey", appKey);
-        headers.set("Authorization", "Bearer " + accessToken);
-
-        HttpEntity<String> entity = new HttpEntity<>(headers);
-
-        ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.GET, entity, String.class);
-
-        return new JSONObject(response.getBody());
-    }
-}
