--- investement/src/main/java/com/example/investment/tradingVolume/controller/dto/TradingVolumeDTO.java
@@ -1,11 +1,0 @@
-package com.example.investment.tradingVolume.controller.dto;
-
-public record TradingVolumeDTO(
-        String stockName,
-        String rank,
-        String currentPrice,
-        String totalVolume,
-        String prevVolume,
-        String volumeChangeRate
-) {
-}
